+++ crates/core/src/properties/energy.rs
@@ -9,10 +9,10 @@ use crate::system::System;
 pub struct CoulombEnergy;
 
 impl Property for CoulombEnergy {
-    type Res = f32;
+    type Res = f64;
 
     fn calculate(&self, system: &System, potentials: &Potentials) -> Self::Res {
-        let mut energy = 0 as f32;
+        let mut energy = 0 as f64;
         for (meta, potential) in &potentials.coulombs {
             for (i, j) in &meta.indices {
                 let elem_i = system.elements[*i];
@@ -29,39 +29,6 @@ impl Property for CoulombEnergy {
         }
         energy
     }
-
-    // fn calculate(&self, system: &System, potentials: &Potentials) -> Self::Res {
-    //     potentials
-    //         .coulombs
-    //         .iter()
-    //         .map(|(meta, potential)| {
-    //             // first calculate the total energy due to self-interaction
-    //             let self_energy: f32 = system
-    //                 .element_counts()
-    //                 .iter()
-    //                 .map(|(elem, count)| potential.energy_self(elem.charge()) * (*count as f32))
-    //                 .sum();
-
-    //             let _sum: f32 = meta
-    //                 .indices
-    //                 .iter()
-    //                 .map(|(i, j)| {
-    //                     let elem_i = system.elements[*i];
-    //                     let elem_j = system.elements[*j];
-    //                     let pos_i = system.positions[*i];
-    //                     let pos_j = system.positions[*j];
-    //                     let r = system.cell().distance(&pos_i, &pos_j);
-    //                     if meta.cutoff > r {
-    //                         potential.energy(elem_i.charge(), elem_j.charge(), r)
-    //                     } else {
-    //                         0 as f32
-    //                     }
-    //                 })
-    //                 .sum();
-    //             _sum - self_energy
-    //         })
-    //         .sum()
-    // }
 }
 
 /// Potential energy due to pairwise potentials.
@@ -69,10 +36,10 @@ impl Property for CoulombEnergy {
 pub struct PairEnergy;
 
 impl Property for PairEnergy {
-    type Res = f32;
+    type Res = f64;
 
     fn calculate(&self, system: &System, potentials: &Potentials) -> Self::Res {
-        let mut energy = 0 as f32;
+        let mut energy = 0 as f64;
         for (meta, potential) in &potentials.pairs {
             for (i, j) in &meta.indices {
                 let pos_i = system.positions[*i];
@@ -85,30 +52,6 @@ impl Property for PairEnergy {
         }
         energy
     }
-
-    // fn calculate(&self, system: &System, potentials: &Potentials) -> Self::Res {
-    //     potentials
-    //         .pairs
-    //         .iter()
-    //         .map(|(meta, potential)| {
-    //             let _sum: f32 = meta
-    //                 .indices
-    //                 .iter()
-    //                 .map(|(i, j)| {
-    //                     let pos_i = system.positions[*i];
-    //                     let pos_j = system.positions[*j];
-    //                     let r = system.cell().distance(&pos_i, &pos_j);
-    //                     if meta.cutoff > r {
-    //                         potential.energy(r)
-    //                     } else {
-    //                         0.0 as f32
-    //                     }
-    //                 })
-    //                 .sum();
-    //             _sum
-    //         })
-    //         .sum()
-    // }
 }
 
 /// Potential energy of the whole system.
@@ -116,7 +59,7 @@ impl Property for PairEnergy {
 pub struct PotentialEnergy;
 
 impl Property for PotentialEnergy {
-    type Res = f32;
+    type Res = f64;
 
     fn calculate(&self, system: &System, potentials: &Potentials) -> Self::Res {
         let coulomb_energy = CoulombEnergy.calculate(system, potentials);
@@ -130,10 +73,10 @@ impl Property for PotentialEnergy {
 pub struct KineticEnergy;
 
 impl IntrinsicProperty for KineticEnergy {
-    type Res = f32;
+    type Res = f64;
 
     fn calculate_intrinsic(&self, system: &System) -> <Self as IntrinsicProperty>::Res {
-        let kinetic_energy: f32 = system
+        let kinetic_energy: f64 = system
             .elements
             .iter()
             .zip(system.velocities.iter())
@@ -148,7 +91,7 @@ impl IntrinsicProperty for KineticEnergy {
 pub struct TotalEnergy;
 
 impl Property for TotalEnergy {
-    type Res = f32;
+    type Res = f64;
 
     fn calculate(&self, system: &System, potentials: &Potentials) -> Self::Res {
         let kinetic = KineticEnergy.calculate(system, potentials);
